Support postgres < 9.6
Postgres < 9.6 does not have the setting `idle_in_transaction_session_timeout`. Attempting to always set `idle_in_transaction_session_timeout = 0` causes all migrations to fail on Postgres 9.5.

This PR suggests attempting it and then restarting the transaction if it is unable to set the setting.

I am not a Go expert or very familiar with this library so this may not be the _right_ way to do it, but it is a way that works for our purposes.